@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { HttpInterceptor, HttpRequest, HttpHandler, HttpEvent, HttpErrorResponse } from '@angular/common/http';
-import { Observable, throwError, BehaviorSubject } from 'rxjs';
+import { Observable, BehaviorSubject, throwError } from 'rxjs';
 import { AuthService } from './auth/shared/auth.service';
-import { catchError, switchMap, filter, take } from 'rxjs/operators';
+import { catchError, switchMap, take, filter } from 'rxjs/operators';
 import { LoginResponse } from './auth/login/login-response.payload';
 
 @Injectable({
@@ -11,38 +11,34 @@ import { LoginResponse } from './auth/login/login-response.payload';
 export class TokenInterceptor implements HttpInterceptor {
 
     isTokenRefreshing = false;
-    // we will assign value of refresh token to it
     refreshTokenSubject: BehaviorSubject<any> = new BehaviorSubject(null);
+
     constructor(public authService: AuthService) { }
 
-    intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-        const jwtToken = this.authService.getJwtToken();
+    intercept(req: HttpRequest<any>, next: HttpHandler):
+        Observable<HttpEvent<any>> {
 
         if (req.url.indexOf('refresh') !== -1 || req.url.indexOf('login') !== -1) {
             return next.handle(req);
         }
+        const jwtToken = this.authService.getJwtToken();
 
-        // add token to header
         if (jwtToken) {
-            this.addToken(req, jwtToken);
+            return next.handle(this.addToken(req, jwtToken)).pipe(catchError(error => {
+                if (error instanceof HttpErrorResponse
+                    && error.status === 403) {
+                    return this.handleAuthErrors(req, next);
+                } else {
+                    return throwError(error);
+                }
+            }));
         }
-        // catch 403 forbidden error for call refresh token
-        return next.handle(req).pipe(catchError(error => {
-            if (error instanceof HttpErrorResponse && error.status === 403) {
-                return this.handleAuthErrors(req, next);
-            } else {
-                return throwError(error);
-            }
-        }));
+        return next.handle(req);
+
     }
 
-    addToken(req: HttpRequest<any>, jwtToken: any) {
-        return req.clone({
-            headers:req.headers.set('Authorization', 'Bearer ' + jwtToken)
-        });
-    }
-
-    private handleAuthErrors(req: HttpRequest<any>, next: HttpHandler) : Observable<HttpEvent<any>> {
+    private handleAuthErrors(req: HttpRequest<any>, next: HttpHandler)
+        : Observable<HttpEvent<any>> {
         if (!this.isTokenRefreshing) {
             this.isTokenRefreshing = true;
             this.refreshTokenSubject.next(null);
@@ -52,20 +48,27 @@ export class TokenInterceptor implements HttpInterceptor {
                     this.isTokenRefreshing = false;
                     this.refreshTokenSubject
                         .next(refreshTokenResponse.authenticationToken);
-                    return next.handle(this.addToken(req, refreshTokenResponse.authenticationToken));
+                    return next.handle(this.addToken(req,
+                        refreshTokenResponse.authenticationToken));
                 })
             )
-        } 
-        
-        else {
+        } else {
             return this.refreshTokenSubject.pipe(
                 filter(result => result !== null),
                 take(1),
                 switchMap((res) => {
-                    return next.handle(this.addToken(req, this.authService.getJwtToken()))
+                    return next.handle(this.addToken(req,
+                        this.authService.getJwtToken()))
                 })
             );
         }
+    }
+
+    addToken(req: HttpRequest<any>, jwtToken: any) {
+        return req.clone({
+            headers: req.headers.set('Authorization',
+                'Bearer ' + jwtToken)
+        });
     }
 
 }
