@@ -1,9 +1,10 @@
-import { AuthService } from './../shared/auth.service';
-import { LoginRequestPayload } from './login-request.payload';
-import { FormGroup, Validators, FormControl } from '@angular/forms';
 import { Component, OnInit } from '@angular/core';
+import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { LoginRequestPayload } from './login-request.payload';
+import { AuthService } from '../shared/auth.service';
 import { Router, ActivatedRoute } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
+import { throwError } from 'rxjs';
 
 @Component({
   selector: 'app-login',
@@ -17,30 +18,28 @@ export class LoginComponent implements OnInit {
   registerSuccessMessage: string;
   isError: boolean;
 
-  constructor(private authService: AuthService,
-    private router: Router,
-    private toastr: ToastrService,
-    private activedRoute: ActivatedRoute) {
+  constructor(private authService: AuthService, private activatedRoute: ActivatedRoute,
+    private router: Router, private toastr: ToastrService) {
     this.loginRequestPayload = {
       username: '',
       password: ''
-    }
-
+    };
   }
 
   ngOnInit(): void {
     this.loginForm = new FormGroup({
       username: new FormControl('', Validators.required),
-      password: new FormControl('', Validators.required),
-    })
+      password: new FormControl('', Validators.required)
+    });
 
-    this.activedRoute.queryParams
-    .subscribe(params => {
-      if(params.registered !== undefined && params.registered === 'true') {
-        this.toastr.success('Signup Successful');
-        this.registerSuccessMessage = 'Please Check your inbox for activation email activate your account before you Login!';
-      }
-    })
+    this.activatedRoute.queryParams
+      .subscribe(params => {
+        if (params.registered !== undefined && params.registered === 'true') {
+          this.toastr.success('Signup Successful');
+          this.registerSuccessMessage = 'Please Check your inbox for activation email '
+            + 'activate your account before you Login!';
+        }
+      });
   }
 
   login() {
@@ -48,14 +47,13 @@ export class LoginComponent implements OnInit {
     this.loginRequestPayload.password = this.loginForm.get('password').value;
 
     this.authService.login(this.loginRequestPayload).subscribe(data => {
-      if (data) {
-        this.isError = false;
-        this.router.navigateByUrl('/');
-        this.toastr.success('Login Successful');
-      } else {
-        this.isError = true;
-      }
+      this.isError = false;
+      this.router.navigateByUrl('');
+      this.toastr.success('Login Successful');
+    }, error => {
+      this.isError = true;
+      throwError(error);
     });
   }
-  
+
 }
