@@ -1,11 +1,11 @@
 import { Injectable, Output, EventEmitter } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { SignupRequestPayload } from '../signup/signup-request.payload';
 import { Observable, throwError } from 'rxjs';
 import { LocalStorageService } from 'ngx-webstorage';
 import { LoginRequestPayload } from '../login/login-request.payload';
 import { LoginResponse } from '../login/login-response.payload';
 import { map, tap } from 'rxjs/operators';
+import { SignupRequestPayload } from '../signup/signup-request.payload';
 
 @Injectable({
   providedIn: 'root'
@@ -42,6 +42,10 @@ export class AuthService {
       }));
   }
 
+  getJwtToken() {
+    return this.localStorage.retrieve('authenticationToken');
+  }
+
   refreshToken() {
     return this.httpClient.post<LoginResponse>('http://localhost:8080/api/auth/refresh/token',
       this.refreshTokenPayload)
@@ -55,10 +59,18 @@ export class AuthService {
       }));
   }
 
-  // Utils
-
-  getJwtToken() {
-    return this.localStorage.retrieve('authenticationToken');
+  logout() {
+    this.httpClient.post('http://localhost:8080/api/auth/logout', this.refreshTokenPayload,
+      { responseType: 'text' })
+      .subscribe(data => {
+        console.log(data);
+      }, error => {
+        throwError(error);
+      })
+    this.localStorage.clear('authenticationToken');
+    this.localStorage.clear('username');
+    this.localStorage.clear('refreshToken');
+    this.localStorage.clear('expiresAt');
   }
 
   getUserName() {
